refactor: moved cached classes to nc.base, moved icp procedure to calibrate.py
